R200 exposure and gain discovery support.
Unfortunately, it is no longer true that get_option and get_option_range have no effect on the running device, so their const qualifiers have been removed. This is a breaking change, but more things like this may be required in the future, so best do it now and minimize pain.