Added generic, non-SSSE3 colourspace conversion.
These can be automatically vectorized by the compiler for many platforms.
The target machine is now detected in the Makefile, and NEON vectorization
is selected for arm-linux-gnueabihf.  Other target machines can be
easily handled in a similar way.  If it does not detect a known platform,
SSSE3 support is assumed.