Skip DKMS enablement when kernel is fully patched
In the case of a fully patched kernel, no need to enable DKMS.
If an older, unpatched kernel is installed the user will need
to manually reinstall the librealsense package to patch the kernel.

Added error checking to ensure DKMS is installed before removing.

If DKMS failed to install, remove the DKMS module.