Fix the library location for ROS Debians
The include(GNUInstallDirs) has resulted in the librealsens.so
being placed in lib/x86_64-linux-gnu instead of lib/ like all of
the other ROS libraries. This inconsistency can cause problem
when compiling other ROS programs against the library.

Switching to using Catkin environment variables for ROS builds
as is the standard for ROS.