Build dummy device class and unit test to check config auto-complete feature [FAILING]
Note: the unit test fails on input 2. The way the auto-complete feature currently works is as follows:
for each profile with a wildcard, find a profile the device supports that matches the requested profile, and doesn't contradict the other requests. assume the found profile will lead to a solution and use it to complete the request.
The way we expect input 2 to be processed is for the config object to realize that there is only one supported Infrared profile with fps = 30, which will force the dimensions on both requested streams.
Instead, the config object looks for some depth profile with fps = 30, runs into Depth @480x360x30 first, and then fails due to no infrared profile @ 480x360x30.