examples: control building of the graphical examples.
The graphical examples require a graphical environment, while the
standard examples can be run in a headless system. Separate the two with
a compile switch. If you want to disable building of the graphical
examples, use

    -DBUILD_GRAPHICAL_EXAMPLES=off

in the cmake command line. By default the graphical examples are built
if BUILD_EXAMPLES is selected, so the default behavior does not change.