Generate cmake package config files
By now developers are supposed either to include librealsense into
their workspace or to make it install to a standard location like
/usr/local. Generating cmake package config files helps to
make librealsense relocatable and more suitable for cross-compilation.

Also it helps to detect if the library is missing at configure time.
For that a develper needs to have in her CMakeList.txt the following:

  cmake_minimum_required(VERSION 2.8.11)

  find_package(realsense CONFIG REQUIRED)
  add_executable(hello main.cpp )
  target_include_directories(hello PRIVATE realsense::realsense)
  target_compile_definitions(hello PRIVATE realsense::realsense)
  target_link_libraries(hello LINK_RIVATE realsense::realsense)

No need to hardcode paths to headers and binaries in case the lib
is installed in a non-standard location.

For ROS builds similar config files are generated auto-magically
by catkin. But catkin maintains its own package registry where
non-catkin packages can't find the lib.

Signed-off-by: Dmitry Rozhkov <dmitry.rozhkov@linux.intel.com>