Python Wrapper: Fixed inheritance issue in 'sensor' class
The rs2::sensor class inherits from the rs2::options class. The options
base class provides some methods that are useful to access on sensor object.
Previous implementation was attempting to access these methods by quasi-casting
the methods in the definition of the python bindings for the sensor class.
This wasn't working - see Issue #799.

This commits changes the definition of the sensor binding class by
creating a new 'options' base class and using PyBind11's natural
inheritance methods instead of attempting to cast.