Fix D435i IMU extrinsics
Previously (the rotational part of) _depth_to_imu was applied to the
IMU to get depth coordinates when we should have applied the inverse.
In addition the translational part was negated.

The aligning of the IMU with the depth camera is separated from the
exported-to-user/eeprom-stored extrinsics such that the stored and
default extrinsics now are with respect to the depth aligned IMU.