Expanded the software device API for emulating realsense devices
Issue #4197

This adds the ability for the user with access to the `software_device` and
`software_sensor` classes to configure and set the `camera_info`, `option`,
and stream profile parameters. This allows for serial number, name, etc
to be written and configured. This also expands the `software_device` to
accept an explicit context argument.