Added a new `emulated_device` group for backend devices
Issue #4197

This commit adds a new group of devices in the backend for emulated
devices. I tried to find a way to make the `playback` devices operate
like emulated devices, but I couldn't find a way that wouldn't break
existing funcionality. I ended up adding a separate emulated devices
group that gets treated similar to hardware devices when used with the
`pipeline` and other context configurations.