Add tracking mask API for T261/T265
Allows the user to "mask" an area of each of the fisheye images to
disallow detections in this region. This will artificially reduce the
available space for tracking and can make results worse. It is
primarily useful if you have a known obstruction of lens and wish to
exclude this zone from tracking.