concurrency: remove arbitrary timeout to avoid infinite loop when empty
The dispatcher was unable tell the difference between "queue empty"
and "timeout while dequeueing" leading to it looping continuously
(taking up 100% CPU in many cases).

By removing the arbitrary timeout, we know that when dequeue returns
false, the queue is empty.  In order to support no timeout, we need to
fully empty the queue with a while loop when we are clearing before
checking _is_alive again.

This also has the effect of draining the queue on stop().