Fix two undefined behavior issues (pipeline/config, elpp)
pipeline/config: initialize playback_loop

We found a UBSan complaint of the value 173 being stored into a boolean in the copy constructor.

third-party/easyloggingpp: work around global init order bug

The validateFileRolling function is called during the static initialization of an el::Storage, but it tries to access the global el::Storage. Clang compiles this into a branch to ud1. Instead, just skip the function call.

PiperOrigin-RevId: 471841322